+++ properties.js
@@ -8,6 +8,7 @@ const infoColor = css.color('#6c757d');
 
 const invalidColor = css.color('#dc3545');
 const validColor = css.color('#198754');
+const audioRecorderColor = css.color('#25D366');
 const recordingColor = css.color('#ff0000');
 
 const darkModeTextColor = lightColor;
@@ -26,6 +27,7 @@ module.exports = {
   darkModeTextColor,
   invalidColor,
   validColor,
+  audioRecorderColor,
   recordingColor,
 
   /* ============================================================= */
@@ -475,7 +477,8 @@ module.exports = {
 
   /* ------------------- Form button layout ---------------------- */
   // From the following choices, select the value that you want to be preselected:
-  // Info as Text: 'bsi-form-info-as-text'
+  // Info as Text below input: 'bsi-form-info-below-input'
+  // Info as Text above input: 'bsi-form-info-as-text'
   // Info as Tooltip: 'bsi-form-info-as-tooltip'
   formTooltipStyle: 'bsi-form-info-as-text',
 
@@ -506,6 +509,10 @@ module.exports = {
   radioGroupLabelText: '',
   radioLabelText: '',
   radioInfoText: '',
+  // From the following choices, select the value that you want to be preselected:
+  // Radios below each other: 'bsi-radio-block'
+  // Radios next to each other: 'bsi-radio-inline'
+  radioOrientationStyle:'bsi-radio-block',
 
   /* ------------------------- Select ---------------------------- */
   // If no text is selected, lorem ipsum sample text is used.
@@ -517,6 +524,14 @@ module.exports = {
   // If no text is selected, lorem ipsum sample text is used.
   textAreaLabelText: '',
   textAreaInfoText: '',
+
+  /* ---------------------- Regex input--------------------------- */
+  // If no text is selected, lorem ipsum sample text is used.
+  regexLabelText: '',
+  regexInfoText: '',
+  // This property is a stringified JSON object which defines the available Regex patterns
+  // example: JSON.stringify({ "Datum": "\\d{2}\\.\\d{2}\\.\\d{4}", "PLZ": "[0-9]{5}", "KFZ": "[A-ZÖÜÄ]{1,3} [A-ZÖÜÄ]{1,2} [1-9]{1}[0-9]{1,3}" })
+  regexPattern: JSON.stringify({}),
 
   /* ----------------------- Error Texts ------------------------- */
   // This text is displayed below a form element if the field is required but has not been filled out.