+++ templates/website/design.js
@@ -76,7 +76,8 @@ module.exports = cx.design
         require("@bsi-cx/design-standard-library-web/content-elements/form/checkbox"),
         require("@bsi-cx/design-standard-library-web/content-elements/form/radio"),
         require("@bsi-cx/design-standard-library-web/content-elements/form/poll"),
-        require("@bsi-cx/design-standard-library-web/content-elements/form/recaptcha")
+        require("@bsi-cx/design-standard-library-web/content-elements/form/recaptcha"),
+        // require("@bsi-cx/design-standard-library-web/content-elements/form/regex-input"), !!! set properties.regexPattern first !!!
       )
   )
   .withWebsite(