+++ CHANGELOG.md
@@ -8,9 +8,17 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Hint: You will find the list of compatible BSI versions for each template in the [Compatibility section of the README](https://github.com/bsi-software/bsi-cx-design-master-template-web?tab=readme-ov-file#compatibility).
 
 ## Next
+* Add new if-else-group element
+
+
+## [1.3.9] - 01.10.2025
+
+### Changed
 * Add audio recorder element
 * Add dark mode property
-* Add new if-else-group element
+* Add radioOrientationStyle property
+* Add regexPattern property
+* Change documentation of formTooltipStyle property 
 
 ## [1.3.8] - 03.06.2025